requirements.txt
streamlit
numpy
python-dotenv
psycopg2-binary
langchain
langchain-community
langchain-openai
openai
supabase
langsmith
httpx
apscheduler
pytz
requests